Core: New method FacilityManager.bl.addHosts.
New method addHost take host names as list of string and if there is a pattern made of square
brackets (e.g. "name[00-12]domain") it generates new host names and
then call old method addHosts, so all hosts are created. The logic is written in the class impl.Utils.
RPC: Now when addHosts is called from RPC, the generation of host names
is enabled.