Core: attribute U:D:sshPublicKey can be null
Null in value is allowed in the module for U:D:sshPublicKey.

We still don't have something like "optional" attributes for service. So
for this attribute we still need to allow null value in module therefore
the services can mark it as "required" and GEN script must be ready to
get null in this attr.

For services that really reuires not null ssh key, we temporary use
U:D:sshPublicAdminKey attribute.