Engine:Fix in switching stuck SEND task to ERROR
There are two ways how to switch task which is get stuck in PROCESSING
state to ERROR. One works only for SEND tasks when there is not enough
tasks results after fixed time. Another is more generic and works for
any tasks which stay in PROCESSING or PLANNED for long time.

The one for fixing SEND was set to apply after a tasks which got stuck
for more then 3 hours, while the generic one apply after tasks which is
got stuck for more than 1 hour. So the one for SEND could never apply.

Moreover the one for SEND also switching GEN dependencies to NONE,
therefore it forced generating of new (up to date) data. This is crucial
because otherwise following situation may occur: One task result on a
cluster always get stuck. Propagation maintainer after fixed time switch
SEND task to error and will reschedule it in near future. Unless it also
switches GEN to NONE, engine will propagate the same old data over and
over (because one task result always get stuck, so SEND never ends in
DONE).

This commit sets code for fixing SEND tasks to apply after task which is
got stuck for 2 or more hours and the one for all tasks to apply after 3
hours. Therefore the one for SEND apply before the generic one and
correctly switch SEND to ERROR and GEN dependencies to NONE. The one for
all tasks will solve bigger problems which first one couldn't. This will
fix the problem described above.