Fix security bug in adding existing hosts
You cannot add host if someone already added same host or destination to
another facility. The only exception is when you are manager of the
other facility too. This prevents attacker to create existing
destination and by that force Perun to propagate malicious data to that
destination.

This commit fixes code for checking rights to add hosts, to work in way
described earlier. Old code allows to add any host.

Old code first added the hosts and after that checks if you have rights
to some facility, where this host (or destination with same hostname) is
already used. Because same thread can read uncommitted  changes from DB,
it always found the facility, that you just add the host to (and of
course you are manager of it). Therefore it allows to add any host.