Optimalization of import from external source
ExtSource*:
 - Methods getGroupSubjects now return string with status. It fills subjects through given parameter now.
 - Private methods changed to support change mentioned above.
---

Methods categorizeMembersForLightweightSynchronization and categorizeMembersForSynchronization were changed.
Now it uses map where key is login from needed external source and value is RichMember attached to this login.
Thanks to this is checking if user is already in group faster.
---

Methods to getCandidate were refactored. Now they count hash code of subject obtained from external source.
Hashcodes of subjects are stored in member-group attribute. If it has same value in the next synchronization cycle,
we don't need to update this member's attributes.
---

Communication with external sources SQL, SQLComplex and LDAP was changed.
Now they can send us only modified members of group, when it is possible.
---

UserDuplicateException was created. It is thrown when candidate,
which should be added to group, represents more than one User.