Bona fide attributes
* Creates attribute
urn:perun:user:attribute-def:virt:elixirBonaFideStatus
* This attribute is used for to indicate that a user is a
researcher. If so, it provides URL:
'http://www.ga4gh.org/beacon/bonafide/ver1.0'.

* The decision depends on attribute 'elixirBonaFideStatusREMS',
if it is not empty then the user is a researcher. If it is
empty, then this module searches in user's affiliations. If
any of those affiliations starts with 'faculty@' then the
user is a researcher as well. Otherwise, null value is set.