GetCompleteCandidates feature
* Created two new methods in Entry:
  * vosManager.getCompleteCandidates(vo, attrNames, searchString)
  * vosManager.getCompleteCandidates(vo, group, attrNames, searchString)
* Those methods are used to find MemberCandidate entities that consists
of Candidate, RichUser and Member entities.
* Those entities are used to decide if some candidate has appropriate
user in perun, member in a Vo or member in a group.
* The first method can be called only by VO_ADMIN and it searches for
candidates in all VO's ExtSources.
* The second method can be called by VO_ADMIN or GROUP_ADMIN. When it is
called by GROUP_ADMIN, it searches for candidates only in ExtSources
assigned to given group. Otherwise, it searches in all VO's ExtSources.

* Also were created appropriate methods in BL layer and method for
searching candidates - "getCandidates()" - was made more general. Now it
searches for candidates in given List of ExtSources and it can be
decided if it should return candidates for those users, who are already
in given VO.

* Added new method to UserManagerBl to covert single user to
rich user with attributes by attributes names.