Tool for moving groups
Moves Group within structure of groups in VO with all it's subgroups. Group id or group name together with vo id or vo short name are required fields. Destination group is optional. If destination group is not defined, moving group (and it's subgroups) is moved as first level group in VO.

Indentation corrected.