Consistency for authz for registrar to change pwd
Registrar role is authorized to change password for any user in any
namespace. Before this commit it was authorized to change it only using
login of the user as input parameter. Now, it's possible also using User
object as a parameter.

Now, there should be consistent authorization regardless of the method
used to do the same operation.