Added pattern list of messages in UserVirtualAttributeCollectedFromUserExSource
Method resolveVirtualAttributeValueChange(...) uses messageNeedsResolve(...) to decide if the generated message has effect on the attribute and new messages have to be generated. 
Patterns to match message are returned in List from getPatternsForMatch(). This method can be further overridden and the list of patterns can be extended in this way. (This can help if we have dependencies on more attributes)