Change of return values of Attribute Cache
Attribute cache returned in almost every method AttributeHolders object instead of AttributeDefinition/Attribute object.
Its changed, so get methods now return same type of object as it is written in the return clause.

Methods for transformation of the list of AttributeHolders to the list of AttributeDefinitions/Attributes were added to BeansUtils.

One test in CacheManagerTest.java was changed because it was dependent on the order in the list and it shouldn't.