RPC: added check of max length of user titles
* Problem: when creating user, there is no check of given values in
titleAfter and titleBefore but the DB limit is 40. So, when longer
titles are given, the InternalErrorException is risen without
reasonable message.

* Solution: added check that in case of titleAfter/titleBefore being
longer than 40 characters rise an MaxSizeExceededException with proper
message. This check is added in method updateNameTitles in usersManager
and in methods createMember in members manager which take a Candidate
as a parameter.