Membership expiration extension fix
There was an error introduced in
0a6a4f8cd1dbfc0b17b77c5048e846a6d963190c which broke expiration
extension with fixed date (i.e. account is extended to fixed date in the
future).

When the extension date were in next year (not current year) and user
was within the grace period, which means he should get one extra year in
his extension, he doesn't get the extra year. Which usually means he got
the same expiration date which he already had.

Problem is hidden in method extendCalendarByStaticDate which have a side
effect of changing the date from parameter to the date of the extension.
This will caused that "gracePeriodCalendar" date computed by
extendGracePeriodCalendar method is prolonged by one extra year, which
shouldn't happen.

For example today is 1.12.2000, extension date is 1.2. and grace period
is 3 months. extendCalendarByStaticDate changes the date in parameter to
1.2.2001 and extendGracePeriodCalendar sets gracePeriodCalendar to
1.11.2001 (that is 1.2.2002 (the extra year which shouldn't be there)
subtract 3 months grace period).

As a result of this extendGracePeriodCalendar seems to be OK and
therefore there is no extra year added to extension date is is should
be.

This commit fixes the addition of one extra year to
extendGracePeriodCalendar which will fix the described issue.