Fix email validation confirmation message
The email address in the message was prefixed by string "safe: ". That
was caused by wrong usage of SafeHtml class. You need to use "asString"
method to get the output. In this case was used toString which returns
result with the "safe: " prefix.

Previous message example:
The email address: safe: "some@email.com"
was verified and set as your preferred email.

New message format:
The email address: some@email.com
was verified and set as your preferred email.