Fix of generating inverse strong dependencies
* Problem: the inverse strong dependencies were not generated properly
because of a typo. There was a wrong object given to the method
responsible for generating iverse dependencies. As a result, the
inverseDependencies map contained only empty sets for each attribute.
This problem could cause some dependencies not being checked while
setting an attribute's value.

* Solution was to give the method for generating the inverse
dependencies the correct object - strongDependencies.