Fix of attribute module dependencies removal
* Problem: during the deletion of an attribute, the old implementation
would always delete the module dependencies from relevant maps
regardless possible failure of the DB removal. If the DB operation would
fail, the dependencies would be removed but the attribute would not
still be deleted. Next attempt of the attribute removal would fail due
to NullPointerException because the dependencies would no longer contain
the attribute's data.
* Fix: move the dependencies removal after the DB operation. That way,
the dependencies are removed only if the DB operation succeeds.