Skipping of value check during dependencies check
* Problem: during dependencies check, the current implementation is
trying to check value of all attributes that are dependant on given
attributes. Even though the attribute doesn't need to implement methods
for value checks (semantic, syntax). Because of that it needs to
calculate its value. This can be slow and it is useless.
* Solution: there is a new annotation for modules of virtual attributes
named SkipValueCheckDuringDependencyCheck. This annotation is used for all
virtual attribute modules that don't check theirs values. During the
dependencies check, the method now skips all attributes that have a
module which is implementing this new interface.