Fix wrong behavior for unixGID when adding group to resource
 - Unix GID wasn't filled for groups, which already had unixGroupName set,
during their assignment to the resource (where both attributes were required).
 - The problem was in group-resource:virt:unixGID module, which didn't react
on correct exception (WrongRefferenceAttributeValueException), but rather on
former WrongAttributeValue (which is now thrown only for the wrong syntax, but
empty values are allowed -> hence attribute was not filled).
 - Now we react on the correct exception and fill unixGID for the groups, which
already have unix group name in the namespace defined by the resource.