Group structure sync - resources
* Implemented login responsible for automatic assignment of resources
for certain group trees in the group structure synchronization.
* Created a new attribute for group structure synchronization named
groupStructureResources. This attribute is of the map type and keys
represents id of a resource which should be assigned, and values are
group logins, seperated with commas. The resource is assigned to group
with the login and also to all of its subgroups.