Roles - default politics file
* With the old implementation, it was difficult to release new versions
of perun with new policits. All configurations would have to be changed
and there was a lot of places to do a manual error. Even right now, we
don't plan to have different politics on different instances.
* Created a default politics cofiguration file
`perun-base/src/main/resources/perun-roles.yml` which is packed into the
war archive and will be used in production and even in tests.
* Created a test configuration file `test-roles.yml` which contains
politics specially created for tests.
* Changed behaviour of loading of the politics. On production, the
default politics configuration is used and also, the configuration file
at `/etc/perun/perun-roles.yml` can be used to override some politics.
This file is ignored during tests.