Perun core - removed SUSPEND status
* With the old implemetation of the member suspension it was difficult
to determine in which state the object was before. It wasn't also
possible to specify a reason why the member has been suspended.
Therefore, a new implementation using bans has been created.
* SUSPEND status has been completely removed and it has been replaced
with new implementation of VoBans.
* The old method `suspendMemberTo` now creates ban objects, but it has
been marked as deprecated and one should use the newly created methods
that work directly with bans.
* Updated isSuspended attribute's module to read the information from
bans.
* Created new methods that work with vo bans.
* New methods added to the OpenApi.
* In the old gui it is not possible to set the suspend status anymore.
* In CLI, removed suspension info from member and adjusted comments.