Core - dont remove non-required u-f attributes
* The old implementation of deleteMember was trying to find out, which
attributes for corresponding user will not be required anymore, after the
member deletion. This is not something we want. We want to persist u-f
attributes, even if they are not required, because the user might return
to the facility and we want him to get the same attributes. Also, if the
member was at first manually removed from his groups, this would not
happen and all u-f would remain set.