Group sync - assing groups to resource at once
* The old implementation of assigning resources was trying to assing
groups one by one, so it could skip the ones that failed for some
reason. This could have much worse performance than assigning all groups
at once, because that way, the member attributes are checked only once.
* Also changes methods for assigning groups to resource so they don't
throw GroupAlreadyAssignedException, since we do not want the assingment
to fail if some of the groups is already assigned (which is what we
wanted to do, so it is no problem).
* Also, changed the input parameter of the assingGroupsToResource from
List<Group> to Iterable<Group> because that way, it can accept groups in
more variable ways and there it actually no reason, why should we force
a List.