Core - Member mapper rework
* The old implementation of MEMBER_MAPPER could be used to load also
member's group statuses. But, it the query didn't return these columns,
it used to throw an Exception. This exception has been catched and the
mapper just skipped the group-specific columns. This caused a
performance issues because exception handling has a high cost
(especially when it was thrown for every row).
* The MEMBER_MAPPER has been split into two mappers - MEMBER_MAPPER
which can be used to load only member fields and
MEMBER_MAPPER_WITH_GROUP which can be used to load also member's group
statuses.