Core - auto extension of members
* The vo's memberExpirationRules attribute can now have two new values
which can be used to automatically extend its members.
* The value of `autoExtensionLastLoginPeriod` has format `^[0-9]+([dmy])$`.
If any user has accessed the Perun at this period at the latest, he
will be extended by the vo's rules, if they have the `period` defined.
* The value of `autoExtensionExtSources` has format of extSource ids
split by commas - `^(\d+)?(,\d+)*$`. If these ids are specified, they
will be used to limit the extSources used to check the last access.