Sponsorship validity
* For deployment:
  * New DB version with changelog
* From now, it is possible to specify the validity of the sponsorship.
The validity is in format yyyy-mm-dd and it is meant to define the
last day when the sponsorship is active. After this day, it will turn
inactive. (NOT PART OF THIS COMMIT)
* Updated existing methods that returned users as sponsors. Now these
methods returns new Sponsor object. It had to be done so it is possible
to also return information about the sponsorship (activeness and
validity). The Sponsor object has the same format as the RichUser, so it
should not cause any incompatibility with our RPC API.
* Created method `updateSponsorshipValidity` which can be used to update
validity of a sponsorship relation.
* Updated policy of the `getSponsors` method so it can be called by
appropriate users.
* BUGFIXES:
  * MembersManagerImpl/getSponsoredMembers - there was a missing
DISTINCT keyword which caused duplicit members being returned if they
had multiple sponsors.
  * MembersManagerMethod/getSponsoredMembersAndTheirSponsors - if no
attrNames were specified, the method used a null which caused
NullPointerException.