Methods for convertion of sponsored users
* Created two new methods that can be used to convert sponsored users.
* Method `convertSponsoredUsers` can be used to create sponsored members
for all sponsored users who are members of the given vo. All sponsors
are kept, if they were not sponsors in the given, the role is set for
them.
* Method `convertSponsoredUsersWithNewSponsor` can also be used for the
conversion. The difference is, that this method sets a given users as a
sponsor of all members. Old sponsors will be not set.
* Once all conversions are done, we have to consider how to remove all
of the user sponsorships. This is not handled by these two methods.