Core - createSponsoredMembersFromCSV
* Users using the method used for creation of multiple users are used to
provide the data in a csv format. Because of that, we have decided to
create a new method which can work with such input.
* Created method createSponsoredMembersFromCSV which takes data
parameter, representing the content of the csv file; and header
parameter, representing the header of the csv file.
* We use the `;` semicolon as a delimiter for both, the body and the
header.
* Required values for now are - `firstname`, `lastname`,
`urn:perun:user:attribute-def:def:preferredMail`.
* Optional values are - `urn:perun:user:attribute-def:def:note`.
* Created new user:def:note attribute. This attribute can be used to
show some note for users in the users overview in GUI.