Core - find optimization
* Implemented optimization of find methods in members manager and users
manager. The key is that there is a search in ues login and it is done
after a table join. However, the search string can limit the records
from ues table before the join. From my testing, this has improved the
execution time of the original SQL from 1.5s to 0.5s.