User anonymization
- New method anonymizeUser() was implemented. It works as replacement for
deleteUser() when it is not possible to delete the user.
- New config property ‘attributesToAnonymize’ was added, it contains list of
attributes names, which can’t be deleted and should be anonymized instead.
- New config property ‘attributesToKeep’ was added, it contains list of
attributes names, which can’t be deleted and should be kept untouched.

Difference from deleteUser():
- User attributes configured in property ‘attributesToAnonymize’ are not
deleted but anonymized instead. The anonymized value of an attribute is
determined from the attr module's method getAnonymizedValue().
- User attributes configured in property ‘attributesToKeep’ are not
deleted but kept untouched instead.