Core - fix getRichUsersFromListOfUsers methods
* There were 2 methods that used to take a list of users as an input and
just fetched for these users their attributes. However, our
deserializers are forbiden to work with UUIDs (for some reason) and thus
when the rich users were missing their UUIDs.
* These methods were rewritten so they just used the ids from the given
list of users and fetch them from DB. This way, we can ensure that the
returned data are up to date and valid.