Core - members pagination
* Created method `getMembersPage` that can be used to get a page of
members of a given vo.
* For now, it is possible to specify the page size, the offset (number
of skipped members), sorting order (ASCENDING/DESCENDING). The possible
columns to sort by are, for now, ID and NAME. However, the name can sort
only in the English collation. To sort in the Czech collation, it is
necessary to have the Czech locale (cs_CZ) installed in the hosting OS.
Then, this collation has to be creation in our DB. And lastly, this
would also need to be done in the test docker conainer, otherwise, the
tests won't pass. These changes are NOT part of this PR.