Fixed subgroup statuses in group members page
- This is a fix for CESNET/perun/#3220 - page of group members with
selected group status could contain members with wrong status if they
have different status in subgroups.
- When getting page of group members with selected group status, all
records with the same group_id and member_id in group_members table are
merged into one row. The group status is evaluated as MIN of all the
group statuses since the member is valid (0) in the group if and only if
he is valid in at least one of the groups. Otherwise, he is expired (1).
- Using group members mapping so membership types can be set to returned
members.
- Added test and modified existing tests to check membership types of
members.