fix(synchronizations): don't validate expired members
* During synchronization, the main method categorizes the candidates into multiple lits. One list
represents candidates who were not present in the group before synchronization. Members for these
candidates were always validates, even though these members already existed (just wasn't members of
the group) and were expired in the vo.
* The fix is to call the `updateMemberStatus()` method only validating
the member directly. This method checks, if the user is expired and
should or shouldn't be validated.
* Note: this method is actually called for the member before he is added
to the group, so it is possible, that the call of the `updateMemberStatus`
could be removed . However, we want to make sure that we didn't forget
about some use-cases when adding the member to the group could change
his status.