feat(registrar): VO and group application notifications adjustment
Notifications for VO application are based on toEmail attribute (of given VO), if it is null, preferred email is used to notify all VO admins.
Notifications for Group application are based on toEmail attribute of given group, if it is null, preferred email is used to notify all group admins (if there are any), otherwise VO admins are notified.