feat(core): module studentIdentifiers changed to create and remove userExtSources with attributes
This module for virtual attribute was generating list of
studentIdentifiers for all groups with approprioate attribute where a
user was a valid member.
Now it reacts on attribute messages about adding, removing... member
to a group. According to a message, it either creates userExtSource
with approprate attributes (spuc, sho, epsa) if it not exists already or
it removes the userExtSource if it exists.