fix(registrar): Fix retrieving VO application while initializing registrar
When we were initializing registrar for group applications all
applications were fetched and checked if some of them belong to the
user. This had to be optimized because we do not need all applications,
just the open ones. Therefore, new method was created for that and it is
used in the initialization process instead of the old one.