feat(core): evaluate consents in getData methods
- Added new method evaluateConsents to the BL layer. It returns all
members that a have valid consent for the propagation of the given
service required attributes to the given facility. It also creates an
UNSIGNED consent for users that have no consent with such required
attributes.
- Evaluating was added to getData methods which are called during
propagation.
- The new method evaluateConsents runs in a new transaction because it
is used by getData methods which run in read-only serializable
transactions.
- Locks for users on consent hubs were created to ensure that two
parallel propagations/transactions correctly creates UNSIGNED consents
for users. If it were no locks and two transactions would concurrently
insert the same UNSIGNED consent, the latter committing transaction
would fail on unique constraint in DB.

BREAKING CHANGE: methods servicesManager/getHashedHierarchicalData and
servicesManager/getHashedDataWithGroups are now returning only data
about users with granted consents and creating consents for the users
with insufficient consents (if the consent logic is turned on)