feat(core): changed lifescienceID module
There was a main logic commented inside the method validatePassword. It was done due to
registrations that trigered the functionality even though we want it to use it for different flow
(manual set up of local accounts). Now the logic is uncommented.
However, it is skipped when the actor inside the session is perunRegistrar.