fix(registrar): quick-fix duplicate notifications
- Two duplicate group notifications were sent to group managers when a
user submitted a group application in the following scenario:
1. first the user submitted an application to the group's VO, which was
auto-approved
2. then he submitted the group application less than ~5-6 seconds after
he submitted the VO application
- The reason is that after the VO application was approved, a new
thread was created which slept for 5 seconds and then validated the
member in VO and handled the user's group applications - sent
notifications to the group's managers. If a group application was
submitted in this time window, a notification was sent both in this
method and in the submitApplication method.
- Now, the new thread doesn't sleep for 5 seconds because it is not
needed. It was needed in the async validate method in BL layer because
the transaction that created the member might have not committed yet.
But in Registrar, such a transaction must have committed because
approveApplication doesn't run in a transaction and before creating the
thread we use method getMemberByUser, hence he should already exist in
the DB.
- This is only quick-fix because there is still a time window when this
situation can happen - especially because validate method can take a
few seconds, too.