fix(registrar): fix duplicate notifications
- Two duplicate group notifications were sent to group managers when a
user submitted a group application in the following scenario:

1. first the user submitted an application to the group's VO, which was
auto-approved
2. then he submitted the group application in less than X seconds

- The reason is that after the VO application was approved, a new
thread was created which slept for 5 seconds and then validated the
member in VO and handled the user's group applications - sent
notifications to the group's managers. If a group application was
submitted in this time window (5 seconds + validation time) an
APP_CREATED_VO_ADMIN notification was sent both in this method and in
the second step - submitting the group application.
- #3694 was a quick-fix which removed the thread's sleep, but there was
still a big enough time window when this bug could happen.
- Now, a new table created_app_notifications_sent is introduced to save
the information whether APP_CREATED_VO_ADMIN notification was already
sent.
- The notification is sent always in the second step (submitting group
application) if the user has a required state.
- The notification is sent in the first step only if it wasn't sent yet
and the user has a required state. There is no need for an additional
synchronization (like checking if it is being sent in a concurrent
thread), because the first step sends the notification only for
committed applications, and they are committed together with
information whether the notification was sent.

BREAKING CHANGE: update DB