feat: 🎸 Cascade to parent gr. when deciding gr. TO recipients
When sending out notification to managers or configured TO recipients
for groups, try to lookup recipients in group hierarchy instead of using
the VO level directly. This means that if the group does not have "TO"
set and there is noone to send notification to in corresponding role,
the lookup will be tried one level higher in the group hierarchy,
cascading to the top level group and only then to VO level.

DEPLOYMENT NOTE: Changed behaviour might cause sending notifications to
managers or configured TO recipients in parent group rather than to VO.